Add file operation interface to NetCDF C++ API
Created an interface for calling the NetCDF C++ API from Fortran.
Currently, only file open and close operations are supported.

This is the first step toward a full Fortran-to-NetCDF C++ interface,
which will expose all NetCDF C++ features to Fortran. This enhancement
addresses the limitation of the NetCDF Fortran API, which does not support
reading or writing NetCDF string variables.