Improved f_c_string code safety.
The original f_c_string code was extremely unsafe as it exposed raw C pointers. The new impl utilizes RAII as much as fortran allows and returns error codes if C memory is overwritten of if unallocated memory is accessed. It is still not safe, but it is better than it was. I also added unit tests for this code.